requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.4.0
psycopg2==2.9.5
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
